requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.6.15
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
decorator==5.2.1
edge-tts==7.0.2
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-genai==1.21.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.2.1
multidict==6.5.0
numpy==2.3.0
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
pillow==11.2.1
playwright==1.52.0
playwright-stealth==2.0.0
praw==7.8.1
prawcore==2.4.0
proglog==0.1.12
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
pyparsing==3.2.3
pyperclip==1.9.0
PySocks==1.7.1
pysrt==1.1.2
python-dotenv==1.1.0
pyttsx3==2.98
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
srt==3.5.3
tabulate==0.9.0
tenacity==8.5.0
tiktok-uploader==1.1.1
toml==0.10.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
update-checker==0.18.0
uritemplate==4.2.0
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
